(maint) move installers to a section by themselves
