+++ src/chocolatey/chocolatey.csproj
@@ -108,7 +108,7 @@
     <Compile Include="infrastructure.app\configuration\PackagesConfigFileSettings.cs" />
     <Compile Include="infrastructure.app\domain\ConfigCommandType.cs" />
     <Compile Include="infrastructure.app\domain\FeatureCommandType.cs" />
-    <Compile Include="infrastructure.app\domain\InstallerBase.cs" />
+    <Compile Include="infrastructure.app\domain\installers\InstallerBase.cs" />
     <Compile Include="infrastructure.app\domain\PackageFile.cs" />
     <Compile Include="infrastructure.app\domain\PackageFiles.cs" />
     <Compile Include="infrastructure.app\domain\PinCommandType.cs" />
@@ -140,14 +140,14 @@
     <Compile Include="infrastructure.app\nuget\NugetEncryptionUtility.cs" />
     <Compile Include="infrastructure.app\services\ITemplateService.cs" />
     <Compile Include="infrastructure.app\domain\ChocolateyPackageInformation.cs" />
-    <Compile Include="infrastructure.app\domain\CustomInstaller.cs" />
-    <Compile Include="infrastructure.app\domain\IInstaller.cs" />
-    <Compile Include="infrastructure.app\domain\InnoSetupInstaller.cs" />
-    <Compile Include="infrastructure.app\domain\InstallShieldInstaller.cs" />
+    <Compile Include="infrastructure.app\domain\installers\CustomInstaller.cs" />
+    <Compile Include="infrastructure.app\domain\installers\IInstaller.cs" />
+    <Compile Include="infrastructure.app\domain\installers\InnoSetupInstaller.cs" />
+    <Compile Include="infrastructure.app\domain\installers\InstallShieldInstaller.cs" />
     <Compile Include="infrastructure.app\domain\InstallerType.cs" />
     <Compile Include="infrastructure.app\domain\InstallTokens.cs" />
-    <Compile Include="infrastructure.app\domain\MsiInstaller.cs" />
-    <Compile Include="infrastructure.app\domain\NsisInstaller.cs" />
+    <Compile Include="infrastructure.app\domain\installers\MsiInstaller.cs" />
+    <Compile Include="infrastructure.app\domain\installers\NsisInstaller.cs" />
     <Compile Include="infrastructure.app\domain\RegistryApplicationKey.cs" />
     <Compile Include="infrastructure.app\nuget\ChocolateyLocalPackageRepository.cs" />
     <Compile Include="infrastructure.app\nuget\ChocolateyNugetLogger.cs" />
